+++ packages/authorization/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nodeboot/authorization
 
+## 1.3.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+    -   @nodeboot/engine@1.6.3
+
 ## 1.3.3
 
 ### Patch Changes