+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/config
 
+## 1.3.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+
 ## 1.3.3
 
 ### Patch Changes