+++ serverless/lambda-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nodeboot/lambda-server
 
+## 1.2.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+    -   @nodeboot/core@1.15.3
+    -   @nodeboot/engine@1.6.3
+
 ## 1.2.2
 
 ### Patch Changes