+++ starters/actuator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nodeboot/starter-actuator
 
+## 1.7.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+    -   @nodeboot/config@1.3.4
+    -   @nodeboot/engine@1.6.3
+
 ## 1.7.3
 
 ### Patch Changes