+++ starters/backstage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nodeboot/starter-http
 
+## 1.3.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+    -   @nodeboot/core@1.15.3
+
 ## 1.3.3
 
 ### Patch Changes