+++ starters/firebase/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nodeboot/starter-firebase
 
+## 1.2.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+    -   @nodeboot/config@1.3.4
+    -   @nodeboot/core@1.15.3
+    -   @nodeboot/engine@1.6.3
+
 ## 1.2.3
 
 ### Patch Changes