+++ starters/openapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nodeboot/starter-openapi
 
+## 2.4.5
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@2.1.0
+    -   @nodeboot/engine@1.6.3
+
 ## 2.4.4
 
 ### Patch Changes